Fixed 2 bugs: 1. Deselecting features previously failed to remove entries from the confirmed table.    This was because I had added an <input> element, which meant a naive comparison    of the html text would fail a match. 2. Filtering on the Database without specifying any features caused the resulting list    to be empty. That is due to an attempt to zip non-empty records with an empty    features list.
Added feature to select and deselect records in discrete tables for records that
are searched-for in the search boxes.

Changed the text of the sample crossfilter tab to make it very clear to the user
that no records were found in the prior database search. The Download button will
also be disabled under that scenario.