Refactored names in views.py to better reflect the work with Features (instead of parameters). Refactored the code to more cleanly process database queries with static fields.
Major changes to the way data is shipped to the client. In this version
ALL features data are shipped to the client regardless of which features
are needed for crossfiltering. This is unlikely to be the version we
will settle on.

Changes to the client include the way the CSV file is generated, which
gives us more control over the order of the columns, and a cleaner way
to eliminate multiple rows sharing the same Zenodo ID.